Escrito todos os testes do mapper

+++ src/database/Mapper.java
@@ -5,10 +5,15 @@ import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
 import java.lang.annotation.ElementType;
+import java.lang.annotation.Repeatable;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -34,6 +39,12 @@ public class Mapper {
 		int size() default -1;
 	}
 	
+	@Retention(RetentionPolicy.RUNTIME)
+	@Target(ElementType.TYPE)
+	public @interface UseTables {
+		String[] value();
+	}
+	
 	private Connection connection;
 	
 	/**
@@ -42,21 +53,6 @@ public class Mapper {
 	 */
 	public Mapper (Connection connection) {
 		this.connection = connection;
-	}
-	
-	/**
-	 * Interface que permite compatibilidade com funcoes de mapeamento
-	 * Classes que implementem essa interface devem garantir que a interface tenha acesso a seus campos, ou seja, todos os campos devem ter no maximo acesso default, caso estejam no package database, e public caso contrario
-	 */
-	public static interface Managed {
-		/**
-		 * @return: Lista com nomes de todas as tabelas que armazenem as informacoes contidas no objeto
-		 */
-		public List<String> getTableDependancies ();
-		/**
-		 * @return: Nova instancia de objeto a ser mapeado
-		 */
-		public Object newInstance();
 	}
 	
 	/**
@@ -168,12 +164,133 @@ public class Mapper {
 		int i = 0;
 		PreparedStatement statement = this.connection.prepareStatement(query);
 		
-		Iterator<Object> it = parameters.iterator();
-		while (it.hasNext()) {
-			statement.setObject(++i, it.next());
+		if (parameters != null) {
+			Iterator<Object> it = parameters.iterator();
+			while (it.hasNext()) {
+				statement.setObject(++i, it.next());
+			}
 		}
 		
 		return statement;
+	}
+	
+	private static String[] getDeclaredTables (Class<?> c) throws IllegalArgumentException {
+		UseTables declTables = c.getDeclaredAnnotation(UseTables.class);
+		if (declTables == null || declTables.value().length == 0) {
+			throw new IllegalArgumentException("Class "+c.getName()+" has no declared tables");
+		}
+		return declTables.value();
+	}
+	
+	private static Object retrieveData (ResultSet resultSet, String columnLabel, Class<?> returnType) throws SQLException, IllegalArgumentException {
+		Object data = resultSet.getObject(columnLabel);
+		String dataType = data.getClass().getSimpleName();
+		String returnTypeName = returnType.getSimpleName();
+		
+		switch (dataType) {
+			case "Integer":
+				Integer intVal = (Integer)data;
+				switch (returnTypeName) {
+					case "byte":
+					case "Byte":
+						data = new Byte(intVal.byteValue());
+					break;
+					
+					case "short":
+					case "Short":
+						data = new Short(intVal.shortValue());
+					break;
+					
+					case "Integer":
+					case "int":
+					break;
+					
+					case "long":
+					case "Long":
+						data = new Long(intVal.longValue());
+					break;
+					
+					default:
+						throw new IllegalArgumentException("Cannot convert type "+dataType+" to "+returnTypeName);
+					
+				}
+			break;
+			
+			case "Long":
+				Long longVal = (Long)data;
+				switch (returnTypeName) {
+					case "byte":
+					case "Byte":
+						data = new Byte(longVal.byteValue());
+					break;
+					
+					case "short":
+					case "Short":
+						data = new Short(longVal.shortValue());
+					break;
+					
+					case "Integer":
+					case "int":
+						data = new Integer(longVal.intValue());
+					break;
+					
+					case "long":
+					case "Long":
+					break;
+					
+					default:
+						throw new IllegalArgumentException("Cannot convert type "+dataType+" to "+returnTypeName);
+					
+				}
+			break;
+			
+			case "Float":
+				Float floatVal = (Float)data;
+				switch (returnTypeName) {
+					case "float":
+					case "Float":
+					break;
+					
+					case "double":
+					case "Double":
+						data = new Double(floatVal.doubleValue());
+					break;
+					
+					default:
+						throw new IllegalArgumentException("Cannot convert type "+dataType+" to "+returnTypeName);
+				}
+			break;
+			
+			case "Double":
+				Double doubleVal = (Double)data;
+				switch (returnTypeName) {
+					case "float":
+					case "Float":
+						data = new Float(doubleVal.floatValue());
+					break;
+					
+					case "double":
+					case "Double":
+					break;
+					
+					default:
+						throw new IllegalArgumentException("Cannot convert type "+dataType+" to "+returnTypeName);
+				}
+			break;
+			
+			case "String":
+			case "Date":
+			case "Char":
+			case "char":
+			case "Boolean":
+			case "boolean":
+				if (!dataType.toUpperCase().equals(returnTypeName.toUpperCase())) {
+					throw new IllegalArgumentException("Cannot convert type "+dataType+" to "+returnTypeName);
+				}
+			break;
+		}
+		
+		return data;
 	}
 	
 	/**
@@ -182,48 +299,56 @@ public class Mapper {
 	 * @throws SQLException
 	 * @throws IllegalArgumentException: Caso tabela nao possua campos ou algum dos campos nao possa ser mapeado
 	 */
-	public void create (Class<? extends Mapper.Managed> table) throws SQLException, IllegalArgumentException {
+	public void create (Class<?> table) throws SQLException, IllegalArgumentException {
 		
 		StringBuilder query = new StringBuilder();
 		StringBuilder primaryKeys = new StringBuilder();
 		Field[] fields = table.getDeclaredFields();
-		ForeignKey fk;
+		String[] tables = Mapper.getDeclaredTables(table);
+		ForeignKey fk = null;
 		
+		//check consistency
 		if (fields.length == 0) {
-			throw new IllegalArgumentException("Class "+table.getSimpleName()+" has no fields");
+			throw new IllegalArgumentException("Class "+table.getName()+" has no fields");
+		}
+		if (tables.length != 1) {
+			System.out.println(tables.length);
+			throw new IllegalArgumentException("Class "+table.getName()+" must only use one table to be created");
 		}
 		
 		query.append("CREATE TABLE ");
-		query.append(table.getSimpleName());
+		query.append(tables[0]);
 		query.append(" (\n");
 		
 		primaryKeys.append("PRIMARY KEY(");
-		
+	
 		for (Field field : fields) {
-			query.append("\t");
-			query.append(field.getName());
-			query.append(' ');
-			query.append(Mapper.mapToSqlType(field));
-			query.append(",\n");
-			
-			//if annotated as primary key add to primary keys
-			if (field.getDeclaredAnnotation(PrimaryKey.class) != null) {
-				primaryKeys.append(field.getName());
-				primaryKeys.append(", ");
-			}
-			
-			//if annoted as foreign key add constraint
-			if ( (fk = field.getDeclaredAnnotation(ForeignKey.class)) != null) {
-				query.append("\tFOREIGN KEY(");
+			if (!Modifier.isStatic(field.getModifiers())) {
+				query.append("\t");
 				query.append(field.getName());
-				query.append(") REFERENCES ");
-				query.append(fk.references());
-				query.append("(");
-				if (fk.on().isEmpty())
+				query.append(' ');
+				query.append(Mapper.mapToSqlType(field));
+				query.append(",\n");
+				
+				//if annotated as primary key add to primary keys
+				if (field.getDeclaredAnnotation(PrimaryKey.class) != null) {
+					primaryKeys.append(field.getName());
+					primaryKeys.append(", ");
+				}
+				
+				//if annoted as foreign key add constraint
+				if ( (fk = field.getDeclaredAnnotation(ForeignKey.class)) != null) {
+					query.append("\tFOREIGN KEY(");
 					query.append(field.getName());
-				else
-					query.append(fk.on());
-				query.append("),\n");
+					query.append(") REFERENCES ");
+					query.append(fk.references());
+					query.append("(");
+					if (fk.on().isEmpty())
+						query.append(field.getName());
+					else
+						query.append(fk.on());
+					query.append("),\n");
+				}
 			}
 		}
 		//remove trailling characters
@@ -243,29 +368,43 @@ public class Mapper {
 	 * @throws IllegalArgumentException: If the object's class was not correctly created
 	 * @throws SQLException
 	 */
-	public void create(Mapper.Managed object) throws IllegalArgumentException, SQLException {
+	public void create(Object object) throws IllegalArgumentException, SQLException {
 		StringBuilder query;
 		Class<?> objClass = object.getClass();
 		Field[] fields = objClass.getDeclaredFields();
 		ArrayList<Object> values = new ArrayList<Object>(fields.length);
+		String[] tables = Mapper.getDeclaredTables(objClass);
+		int i;
 		
 		//check consistency
 		if (fields.length == 0) {
 			throw new IllegalArgumentException("object has no attributes");
+		}
+		if (tables.length != 1) {
+			throw new IllegalArgumentException("Object of Type "+objClass.getName()+" must only use one table to be inserted");
 		}
 		
 		try {
 			query = new StringBuilder();
 			
 			query.append("INSERT INTO ");
-			query.append(objClass.getSimpleName());
+			query.append(tables[0]);
 			query.append(" VALUES (");
 			
+			//skip static fields
+			i = 0;
+			while (Modifier.isStatic(fields[i].getModifiers()) && i < fields.length) { i++; };
+			if (i == fields.length) {
+				throw new IllegalArgumentException("Class "+objClass.getName()+" only has static fields");
+			}
+			
 			query.append('?');
-			values.add(fields[0].get(object));
-			for (int i = 1; i < fields.length; i++) {
-				query.append(",?");
-				values.add(fields[i].get(object));
+			values.add(fields[i].get(object));
+			for ( i++; i < fields.length; i++) {
+				if (!Modifier.isStatic(fields[i].getModifiers())) {
+					query.append(",?");
+					values.add(fields[i].get(object));
+				}
 			}
 			
 			query.append(')');
@@ -283,76 +422,80 @@ public class Mapper {
 	 * 		SELECT	colum1, colum2, ..., columN
 	 * 		FROM	table1 NATURAL JOIN table2 NATURAL JOIN ... tableN
 	 * 		WHERE	condition1 AND condition2 AND ... conditionN
-	 * @param maxRows: Numero maximo de linhas a serem trazidas, 0 indica que nao ha limite
+	 * @param maxRows: Numero maximo de linhas a serem trazidas, -1 indica que nao ha limite
 	 * @param factory: Fabrica de objetos que instancia objetos a serem mapeados
 	 * @param filter: Filtro obrigatorio
 	 * @param others: Filtros opcionais
 	 */
-	public LinkedList<Object> read(int maxRows, Mapper.Managed factory, Filter filter, Filter... others) throws NoSuchFieldException, IllegalArgumentException, SQLException {
+	public <Type> List<Type> read(int maxRows, Class<Type> factory, Filter... filters) throws IllegalArgumentException, SQLException {
 		
-		Object prop = factory.newInstance();
-		Class<?> outputClass = prop.getClass();
-		Field[] fields = outputClass.getDeclaredFields();
-		List<String> tables = factory.getTableDependancies();
-		
+		Constructor<?> constructor;
+		Field[] fields = factory.getDeclaredFields();
+		String[] tables = Mapper.getDeclaredTables(factory);
 		StringBuilder query = new StringBuilder();
-		ArrayList<Object> params = new ArrayList<Object>(others.length+1);
+		ArrayList<Object> params = null;
 		ResultSet queryResult;
-		Iterator<String> it;
+		int i;
 		
-		LinkedList<Object> rows;
+		List<Type> rows = (maxRows > 0)? new ArrayList<Type>(maxRows) : new LinkedList<Type>();
 		
 		if (fields.length == 0) {
-			throw new IllegalArgumentException("object generated by factory has no attributes");
-		} else if (tables.size() == 0) {
-			throw new IllegalArgumentException("factory has not declared table dependancies");
+			throw new IllegalArgumentException("object generated by factory has no fields");
 		}
-		//checa se todos os filtros estao validos, excecao sera lancada se classe nao possuir algum dos parametros do filtro
-		outputClass.getDeclaredField(filter.getParamFilter());
-		for (Filter f : others) {
-			outputClass.getDeclaredField(f.getParamFilter());
+		
+		//skip static fields
+		i = 0;
+		while (Modifier.isStatic(fields[i].getModifiers()) && i < fields.length) { i++; };
+		if (i == fields.length) {
+			throw new IllegalArgumentException("Class "+factory.getName()+" only has static fields");
 		}
 		
 		//preencher select da query
 		query.append("SELECT\t");
-		query.append(fields[0]);
-		for (int i = 0; i < fields.length; i++) {
-			query.append(",\n\t\t\t");
-			query.append(fields[i].getName());
+		query.append(fields[i].getName());
+		for (i++; i < fields.length; i++) {
+			if (!Modifier.isStatic(fields[i].getModifiers())) {
+				query.append(",\n\t");
+				query.append(fields[i].getName());
+			}
 		}
 		
 		//preencher from da query
-		it = tables.iterator();
 		query.append("\nFROM\t");
-		query.append(it.next());
-		while (it.hasNext()) {
-			query.append(" NATURAL JOIN\n\t\t");
-			query.append(it.next());
+		query.append(tables[0]);
+		for (i = 1; i < tables.length; i++) {
+			query.append(" NATURAL JOIN\n\t");
+			query.append(tables[i]);
 		}
 		
 		//preencher where da query e parametros
-		query.append("\nWHERE\t");
-		query.append(filter.getParamFilter());
-		
-		params.add(filter.getParamValue());
-		
-		for (int i = 0; i < others.length; i++) {
-			query.append(" AND\n\t\t");
-			query.append(others[i].getParamFilter());
+		if (filters.length != 0) {
+			params = new ArrayList<Object>(filters.length);
 			
-			params.add(others[i].getParamValue());
+			query.append("\nWHERE\t");
+			query.append(filters[0].getParamFilter());
+			params.add(filters[0].getParamValue());
+			for (i = 1; i < filters.length; i++) {
+				query.append(" AND\n\t");
+				query.append(filters[i].getParamFilter());
+				params.add(filters[i].getParamValue());
+			}
 		}
 		
 		queryResult = this.generateQuery(query.toString(), params).executeQuery();
-		rows = new LinkedList<Object>();
 		
 		try {
+			constructor = factory.getConstructor();
 			queryResult.next();
 			while (!queryResult.isAfterLast() && maxRows != 0) {
 				//instanciar objeto da linha atual
-				Object obj = factory.newInstance();
+				@SuppressWarnings("unchecked")
+				Type obj = (Type)constructor.newInstance();
 				for (Field field : fields) {
-					field.set(obj, queryResult.getObject(field.getName()));
+					if (!Modifier.isStatic(field.getModifiers())) {
+						Object data = Mapper.retrieveData(queryResult, field.getName(), field.getType());
+						field.set(obj, data);
+					}
 				}
 				
 				rows.add(obj);
@@ -360,8 +503,10 @@ public class Mapper {
 				queryResult.next();
 				maxRows--;
 			}
-		} catch (IllegalAccessException e) {
+		} catch (IllegalAccessException|InstantiationException e) {
 			e.printStackTrace();
+		} catch (InvocationTargetException|NoSuchMethodException e) {
+			throw new IllegalArgumentException("Class "+factory.getName()+" needs default constructor for query to be made");
 		}
 		
 		return rows;
@@ -372,43 +517,60 @@ public class Mapper {
 	 * Objeto sera procurado na base de acordo com suas chaves primarias e entao seus valores serao modificados
 	 * @param object: objeto a ser atualizado com os valores a serem atualizados
 	 */
-	public void update (Mapper.Managed object) throws SQLException {
+	public void update (Object object) throws SQLException {
 		StringBuilder query = new StringBuilder();
 		Class<?> objClass = object.getClass();
 		Field[] fields = objClass.getDeclaredFields();
+		String[] tables = Mapper.getDeclaredTables(objClass);
 		LinkedList<Field> pks = new LinkedList<Field>();
 		ArrayList<Object> params = new ArrayList<Object>(fields.length);
 		Iterator<Field> it;
+		int i;
 		Field field;
 		
+		if (tables.length != 1) {
+			throw new IllegalArgumentException("Object of Type "+objClass.getName()+" must only use one table to be updated");	
+		}
+		
 		query.append("UPDATE ");
-		query.append(objClass.getSimpleName());
+		query.append(tables[0]);
 		query.append(" SET\n");
 		
 		try {
+			
+			//skip static fields
+			i = 0;
+			while (Modifier.isStatic(fields[i].getModifiers()) && i < fields.length) { i++; };
+			if (i == fields.length) {
+				throw new IllegalArgumentException("Class "+objClass.getName()+" only has static fields");
+			}
 		
-			query.append("\n\t\t");
-			query.append(fields[0].getName());
+			query.append("\n\t");
+			query.append(fields[i].getName());
 			query.append("=?");
-			params.add(fields[0].get(object));
-			for (int i = 1; i < fields.length; i++) {
-				if (fields[i].getDeclaredAnnotation(PrimaryKey.class) != null) {
-					pks.add(fields[i]);
-				} else {
-					query.append(",\n\t\t");
-					query.append(fields[i].getName());
-					query.append("=?");
-					params.add(fields[i].get(object));
+			params.add(fields[i].get(object));
+			for (i++; i < fields.length; i++) {
+				if (!Modifier.isStatic(fields[i].getModifiers())) {
+					if (fields[i].getDeclaredAnnotation(PrimaryKey.class) != null) {
+						pks.add(fields[i]);
+					} else {
+						query.append(",\n\t");
+						query.append(fields[i].getName());
+						query.append("=?");
+						params.add(fields[i].get(object));
+					}
 				}
 			}
 			
 			it = pks.iterator();
-			query.append("\nWHERE\n\t\t");
-			query.append(it.next().getName());
+			field = it.next();
+			query.append("\nWHERE\n\t");
+			query.append(field.getName());
 			query.append("=?");
+			params.add(field.get(object));
 			while (it.hasNext()) {
 				field = it.next();
-				query.append(" AND\n\t\t");
+				query.append(" AND\n\t");
 				query.append(field.getName());
 				query.append("=?");
 				params.add(field.get(object));
@@ -426,21 +588,30 @@ public class Mapper {
 	 * @param table: Classe que utiliza a tabela a ser dropada
 	 * @throws SQLException
 	 */
-	public void delete (Class<? extends Mapper.Managed> table) throws SQLException {
-		this.connection.prepareStatement("DROP TABLE "+table.getSimpleName()).executeUpdate();
+	public void delete (Class<?> table) throws SQLException {
+		String[] tables = Mapper.getDeclaredTables(table);
+		if (tables.length != 1) {
+			throw new IllegalArgumentException("Class "+table.getName()+" must only use one table to be dropped");
+		}
+		this.connection.prepareStatement("DROP TABLE "+tables[0]).executeUpdate();
 	}
 	/**
 	 * Exclui objeto da base de dados
 	 * @param object: Objeto a ser excluido
 	 */
-	public void delete (Mapper.Managed object) throws SQLException {
+	public void delete (Object object) throws SQLException {
 		StringBuilder query = new StringBuilder();
 		Class<?> objClass = object.getClass();
 		Field[] fields = objClass.getDeclaredFields();
+		String[] tables = Mapper.getDeclaredTables(objClass);
 		LinkedList<Object> params = new LinkedList<Object>();
 		
+		if (tables.length != 1) {
+			throw new IllegalArgumentException("Object of Type "+objClass.getName()+" must only use one table to be deleted");
+		}
+		
 		query.append("DELETE FROM ");
-		query.append(objClass.getSimpleName());
+		query.append(tables[0]);
 		query.append(" WHERE\n");
 		
 		try {
